style(All): 요구사항 출력에 맞게 스타일 수정

+++ src/main/java/lotto/domain/money/Money.java
@@ -42,7 +42,7 @@ public class Money {
 
 	public void isNumber(String money) {
 		try {
-			Integer.parseInt(money);
+			Long.parseLong(money);
 		} catch (Exception e) {
 			throw new IllegalArgumentException(LOTTO_PURCHASE_AMOUNT_IS_NUMBER);
 		}
